Refactor workflow load stats computation
This commit improves the logic to compute the workflow load 
without requiring to carry a sql query each time new tasks 
are submitting and keeping instead the current state in a 
local map object. 


It also add the ability to detected stalled executions marking
them with a specific UNKNOWN state.